requirements.txt
# Core API dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
requests==2.31.0

# LangChain and OpenAI
langchain==0.1.9
langchain-community==0.0.24
langchain-openai==0.0.8
openai>=1.10.0,<2.0.0
langsmith==0.1.17

# Vector store and embeddings
chromadb==0.4.22
sentence-transformers==2.5.1


# PDF processing
pypdf==3.17.1

# Environment and utilities
python-dotenv==1.0.0

# Voice processing
SpeechRecognition==3.10.1
pyaudio==0.2.14
pydub==0.25.1

# Streamlit and audio recording
streamlit==1.32.0
audio-recorder-streamlit==0.0.10

# Schedule and report generation
schedule==1.2.1
reportlab==4.0.8

# Data processing
pydantic>=2.7.0
pydantic-settings==2.9.1
numpy>=1.24.0
pandas>=2.0.0

# Additional utilities
python-dateutil>=2.8.0

# PDF extraction (used in app)
pdfplumber==0.11.0

# OCR for image-based or handwritten PDFs (used in app_kasa.py)
pytesseract==0.3.10
Pillow==10.3.0

# Word document extraction (used in logic/file_extractor.py)
python-docx==1.1.0

# Excel file extraction (used in logic/file_extractor.py)
openpyxl==3.1.2
